[REFACTOR][IR] Phase out class Integer and class Bool in Attrs and PassConfig
Now that the ffi container machinery (Array, Optional, Map, Variant)
accepts bare int64_t and bool, the Integer/Bool ObjectRef wrappers add
no value in attribute fields, pass-config options, function-attr
flags, and operator-attr-map registries. Every reader paid an extra
.IntValue() / ->value unbox per access for no information gain. This
commit migrates the bulk of those sites at the field-declaration and
call-site level.

Main changes:

- Relax Attrs fields: every `Integer axis;` / `Bool keepdims;` and the
  container forms `Array<Integer>` / `Optional<Array<Integer>>` /
  `Optional<Integer>` / `Optional<Bool>` migrated to the bare type
  (`int64_t` / `bool`). Headers under `include/tvm/relax/attrs/` plus
  builder/Frame and target/VirtualDevice.
- OpAttrMap registry: `set_attr<Bool>("FPurity", Bool(true))` and the
  paired `GetAttrMap<Bool>("FPurity")` readers migrated to `bool`.
- PassContext config: `TVM_REGISTER_PASS_CONFIG_OPTION("...", Bool)`
  and `("...", Integer)` and their `GetConfig<Bool>` / `GetConfig<Integer>`
  readers migrated to bare types.
- Function-attr flags: `GetAttr<Bool>(k, Bool(false)).value()` and
  `GetAttr<Integer>(k, Integer(N)).value().IntValue()` migrated to
  `GetAttr<bool>(k, false).value()` and `GetAttr<int64_t>(k, N).value()`.
- `HasNonzeroAttr` in `ir/attrs.h` dropped its `.IntValue()` unbox.
- Schedule decision arrays (SampleCategorical candidates, perfect-tile
  factors, autobind thread_extents, multi-level-tiling levels, etc.)
  migrated to `Array<int64_t>` / `Optional<int64_t>` since they are
  integer payloads, not IR exprs. This is a virtual-signature change
  on ConcreteScheduleNode::SampleCategorical and related methods.
- `Variant<Bool, Array<String>>` for `LiftTransformParams.shared_transform`
  migrated to `Variant<bool, Array<String>>`.

IR-position `Integer(N)` / `Bool(true|false)` constructors that flow
into PrimExpr / Stmt positions are left alone in this commit; they
are valid as long as `class Integer` and `class Bool` still exist
(both still derive from `IntImm` and share `IntImmNode`). A follow-up
commit will rewrite those to `IntImm(DataType::Int(32), N)` /
`const_true()` / `const_false()` and then delete the two classes.

Build verified clean for CPU target. Smoke-tested Python suites:
relax/test_op_* (130+296 passes), relax/test_transform_legalize_ops,
test_transform_fuse_ops, test_transform_lift_transform_params (56
passes), and tests/python/ir + tests/python/tirx-base (378 passes,
25 skipped).

+++ include/tvm/relax/attrs/manipulate.h
@@ -45,7 +45,7 @@ struct ConcatAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in expand_dims operators */
 struct ExpandDimsAttrs : public BaseAttrsNode {
-  ffi::Array<Integer> axis;
+  ffi::Array<int64_t> axis;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -98,7 +98,7 @@ struct LayoutTransformAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in permute_dims operator */
 struct PermuteDimsAttrs : public BaseAttrsNode {
-  ffi::Optional<ffi::Array<Integer>> axes;
+  ffi::Optional<ffi::Array<int64_t>> axes;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -126,7 +126,7 @@ struct SplitAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in squeeze operators */
 struct SqueezeAttrs : public BaseAttrsNode {
-  ffi::Optional<ffi::Array<Integer>> axis;
+  ffi::Optional<ffi::Array<int64_t>> axis;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -141,7 +141,7 @@ struct SqueezeAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in stack operators */
 struct StackAttrs : public BaseAttrsNode {
-  ffi::Optional<Integer> axis;
+  ffi::Optional<int64_t> axis;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -174,7 +174,7 @@ struct RepeatAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in tile operators */
 struct TileAttrs : public BaseAttrsNode {
-  ffi::Array<Integer> repeats;
+  ffi::Array<int64_t> repeats;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -198,7 +198,7 @@ struct FlipAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in gather_elements operators */
 struct GatherElementsAttrs : public BaseAttrsNode {
-  Integer axis;
+  int64_t axis;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -212,7 +212,7 @@ struct GatherElementsAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in gather_nd operators */
 struct GatherNDAttrs : public BaseAttrsNode {
-  Integer batch_dims;
+  int64_t batch_dims;
 
   static void RegisterReflection() {
     namespace refl = tvm::ffi::reflection;
@@ -252,7 +252,7 @@ struct MeshgridAttrs : public BaseAttrsNode {
 
 /*! \brief Attributes used in scatter_elements operators */
 struct ScatterElementsAttrs : public BaseAttrsNode {
-  Integer axis;
+  int64_t axis;
   ffi::String reduction;
 
   static void RegisterReflection() {
